Overload operators for Pokemon class and confirm correctness
Overloaded comparison operators for the pokemon class so I no longer
have to worry about comparing the correct attribute of the class in my
sorting methods. Also, added a method to confirm the output from my
implementation of the sorting algorithms is correct by using python's
native list.sort() method to generate a correct list to compare the
others to.

CS315 HW1