Add output of sorted pokemon list to file
The project requires I output the newly sorted pokemon to either stdout
or to a file. I stored the sorted lists in a nested dictionary, which
works well with the python yaml library, so I stored the sorted lists to
a yaml file located in the current working directory when the script is
invoked.

CS315 HW1