Add dir for CPE380
In my computer organization course, we are beginning to write "code"
that is the RTL control signals to implement MIPS instructions on a
simple processor architecture. This added dir will help me save this
stuff.

In addition, this commit adds a file containing the RTL signals to
implement a few MIPS instructions on this processor.

CPE380 HW2