Improve Visual Design and Add HTML Tables and Forms
I improved the visual appeal of the page by changing the colors and
adding shadowing to the page menu. I added two new pages: a restaurant
review section and a bug report/feature request section to meet the
requirements of project 3 for my CS316 course which requires html tables
and forms to be added to my web page.

CS316 Project 3