Restructure to support new layout
I am going to be adding a new layout, so I am restructuring the page to
support this. Now, the website is separated into a bunch of different
web pages, each with the same navigaiton bar. Now I need to apply the
css from project 3 to get it to look the same and then add the layouts.

CS316 Project 4