Implement Tree Traversals and Add Wasted Memory Computation
I implemented the last of the tree traversals for the binary search
tree. I am finished with implementing the bst and now I just need to
test my implementation. I also added a member function that computes how
much memory is wasted for the array-based bst because I noticed it was
quite a bit.

CS315 HW2