Create directory to begin project 5 work
Created a project 5 directory with a base template of how I want the
website to look for project 5. This is just what the website looked like
befor I added the thumbnail and the sliding menu (which I didn't care
for but simply added for project requirements).

Also added some code used in computer organization homework.

CS316 Project 5