Add Server Side Scripts
Created a database using mysql that has one table storing quotes from
The Office TV show, and another storing bug report or feature request
information from users.

The user may report a bug or feature request. If they do, they may
also provide their contact information and rate the website. The
contact info and rating is optional. If a report is submitted, it
will be stored in the database.

Created a random The Office quote generator where when the user visits
the page, it queries the database for a random quote from the office
and displays it. The user may also query the database for a random quote
from a specific character. If a quote from that character is found, it
is displayed.