Create initial machine learning (CS460G) commit
Initial commit for machine learning course (CS460G) at UK. This commit
features the beginnings of my code to implement the training and testing
of a decision tree binary classifier on several different data sets.

These data sets were provided by Dr. Brent Harrison at the University of
Kentucky.

So far, I have implemented the loading in of the data sets for the first
part of the assignment, and the discretization of said data

CS460G HW1