Add depth limit and begin pokemon implementation
Added a depth limit to ID3 function so decision trees can't have a
height larger than the limit.

Began implementing pokemon classifier. Started by loading in the data
and getting it in a usable format. The next thing to do is to go through
the ID3 function and make sure it can handle the pokemon data correctly.

CS460G HW1