Complete Pokemon decision tree implementation
Completed the training and testing of pokemon decision tree. I changed
the interface of a couple of methods and added a method or two to
support the pokemon data structure. Luckily, the class I created to
represent the synthetic data worked nicely for the pokemon data as well.

CS460G HW1