Fix colormap creation method to fit data
The colormap created for each file did not fit the scatter plots of the
data because it was not created based on the points around the given
examples in each data set.

Altered the method to actually plot points in the same area of the
feature space as each data set.

CS460G HW1