Implement training on wine data
Implemented the training functions to train a multiple linear regression
model for the wine data. This involves writing a function that performs
gradient descent and a function that computes the value of the partial
derivative of the loss function with respect to the given theta that's
being updated. The training stops after a set number of time steps.

CS460G HW2