Feed forward neural network initial commit
First commit for homework assignment 3: implementing a feed forward
neural network using a multilayer perceptron to classify the MNIST data
set. The data is provided by Dr. Brent Harrison at the University of
Kentucky.

This first commit implements the inital load of training and test data,
the activation function, and the first derivative of the activation
function.

CS460G HW3