Add divergence detection and correction
I have found experimentally that divergence occurs pretty frequently.
Added a divergence detection that will trigger if the previous accuracy
is better than the current accuracy or if the last ten accuracies are
all the same.

If divergence is detected, restart training with a new random
intialization of models.

CS460G HW3